+++ src/lib.rs
@@ -6,5 +6,6 @@ pub use accessor;
 pub use extended_capabilities::ExtendedCapability;
 pub use registers::Registers;
 
+pub mod context;
 pub mod extended_capabilities;
 pub mod registers;